requirements.txt
openai==1.77.0
pydantic==2.11.4
python-dotenv==1.1.0
requests==2.32.3
